refactor: rename mod time to clock
And safer type convertion from i64 (timestamp) to u64.
