+++ src/databases/mysql.rs
@@ -9,8 +9,8 @@ use crate::models::torrent::TorrentListing;
 use crate::models::torrent_file::{DbTorrentAnnounceUrl, DbTorrentFile, DbTorrentInfo, Torrent, TorrentFile};
 use crate::models::tracker_key::TrackerKey;
 use crate::models::user::{User, UserAuthentication, UserCompact, UserProfile};
+use crate::utils::clock::current_time;
 use crate::utils::hex::bytes_to_hex;
-use crate::utils::time::current_time;
 
 pub struct MysqlDatabase {
     pub pool: MySqlPool,
