--- src/utils/time.rs
@@ -1,3 +1,0 @@
-pub fn current_time() -> u64 {
-    chrono::prelude::Utc::now().timestamp() as u64
-}
